FWB-11 등록 요청에 대한 제약 사항 추가 & 커스텀 도메인 추가

+++ src/main/java/kvoting/intern/flowerwebapp/domain/DomainService.java
@@ -4,64 +4,33 @@ import kvoting.intern.flowerwebapp.dict.Dict;
 import kvoting.intern.flowerwebapp.dict.DictService;
 import kvoting.intern.flowerwebapp.dict.registeration.DictReg;
 import kvoting.intern.flowerwebapp.dict.registeration.DictRegService;
-import kvoting.intern.flowerwebapp.domain.registration.request.DomainRegistRequest;
-import kvoting.intern.flowerwebapp.registration.ItemService;
-import kvoting.intern.flowerwebapp.registration.ProcessType;
-import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
+import kvoting.intern.flowerwebapp.item.ItemServiceImpl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashSet;
-
 @Service
-@RequiredArgsConstructor
-@Transactional
-public class DomainService implements ItemService<Domain, DomainRegistRequest> {
-    private final DomainRepository domainRepository;
+public class DomainService extends ItemServiceImpl<Domain> {
     private final DictService dictService;
     private final DictRegService dictRegService;
-    private final ModelMapper modelMapper;
+
+    public DomainService(DomainRepository domainRepository, DictService dictService, DictRegService dictRegService) {
+        super(domainRepository);
+        this.dictService = dictService;
+        this.dictRegService = dictRegService;
+    }
 
     @Transactional(readOnly = true)
     public Page<Domain> getDomainByEngNameContains(String engName, Pageable pageable) {
-        return domainRepository.findByDomainBase_EngNameContains(engName, pageable);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public Domain get(Long id) {
-        return domainRepository.findById(id).orElseThrow(() -> {
-            throw new RuntimeException();
-        });
-    }
-
-    @Override
-    public Domain map(DomainRegistRequest request) {
-        Domain map = modelMapper.map(request, Domain.class);
-        map.setStatus(ProcessType.UNHANDLED);
-        map.setDomainRegs(new HashSet<>());
-        return map;
-    }
-
-    @Override
-    public Domain save(Domain domain) {
-        return domainRepository.save(domain);
-    }
-
-
-    @Override
-    public void setStatus(Domain item, ProcessType type) {
-        item.setStatus(type);
+        return ((DomainRepository) itemRepository).findByDomainBase_NameContains(engName, pageable);
     }
 
     @Override
     public void delete(Domain domain) {
         domain = get(domain.getId());
         for (Dict dict : domain.getDicts()) {
-            if (dict.getDomains().size() == 1) {
+            if (dict.getDomains().size() == 1 && dict.getCustomDomains().size() == 0) {
                 dictService.delete(dict);
                 continue;
             }
@@ -69,12 +38,17 @@ public class DomainService implements ItemService<Domain, DomainRegistRequest> {
             dictService.save(dict);
         }
 
-        domainRepository.flush();
+        itemRepository.flush();
 
         for (DictReg dictReg : domain.getDictRegs()) {
             dictReg.getDomains().remove(domain);
             dictRegService.save(dictReg);
         }
-        domainRepository.delete(domain);
+        itemRepository.delete(domain);
+    }
+
+    @Override
+    public void delete(Long id) {
+        delete(get(id));
     }
 }
